refactor(input): organize modules by control type
