+++ crates/base/src/input/base/blink_cursor.rs
