+++ crates/base/src/input/base/mask_pattern.rs
