--- crates/base/src/input/controls/controls.rs
@@ -1,66 +1,0 @@
-use gpui::{App, Entity, IntoElement, RenderOnce, Window};
-
-use super::{EditorState, InputState, TextareaState};
-
-/// An unstyled single-line text input.
-///
-/// Applications that need a fully styled control can wrap this state with
-/// their own presentation or use `gpui-component::Input`.
-#[derive(IntoElement)]
-pub struct Input {
-    state: Entity<InputState>,
-}
-
-impl Input {
-    pub fn new(state: &Entity<InputState>) -> Self {
-        Self {
-            state: state.clone(),
-        }
-    }
-}
-
-impl RenderOnce for Input {
-    fn render(self, _: &mut Window, _: &mut App) -> impl IntoElement {
-        self.state
-    }
-}
-
-/// An unstyled ordinary multi-line text input.
-#[derive(IntoElement)]
-pub struct Textarea {
-    state: Entity<TextareaState>,
-}
-
-impl Textarea {
-    pub fn new(state: &Entity<TextareaState>) -> Self {
-        Self {
-            state: state.clone(),
-        }
-    }
-}
-
-impl RenderOnce for Textarea {
-    fn render(self, _: &mut Window, _: &mut App) -> impl IntoElement {
-        self.state
-    }
-}
-
-/// An unstyled source-code editor.
-#[derive(IntoElement)]
-pub struct Editor {
-    state: Entity<EditorState>,
-}
-
-impl Editor {
-    pub fn new(state: &Entity<EditorState>) -> Self {
-        Self {
-            state: state.clone(),
-        }
-    }
-}
-
-impl RenderOnce for Editor {
-    fn render(self, _: &mut Window, _: &mut App) -> impl IntoElement {
-        self.state
-    }
-}
