+++ crates/base/src/input/editor/search.rs
